Add notification to Settings Page when local cache is cleared (#221)
* add notification and edit button

* Fix Notification on Settings pages

+++ app/src/pages/SettingsPage.vue
@@ -2,10 +2,27 @@
 import LButton from "@/components/button/LButton.vue";
 import LCard from "@/components/common/LCard.vue";
 import { db, getSocket, isConnected } from "luminary-shared";
+import { useNotificationStore } from "@/stores/notification";
+
+const { addNotification } = useNotificationStore();
 
 const deleteLocalData = async () => {
+    if (!isConnected.value) {
+        return addNotification({
+            title: "Can't clear local cache",
+            description: "You are offline, new data can't be loaded. Wait until you are online.",
+            state: "error",
+        });
+    }
+
     await db.purge();
     getSocket().requestData();
+
+    return addNotification({
+        title: "Local cache cleared",
+        description: "New data is loading from the server, it might take a minute.",
+        state: "success",
+    });
 };
 </script>
 
@@ -19,11 +36,7 @@ const deleteLocalData = async () => {
                 all local data. Depending on the amount of available data on the server, it can take
                 some time before all data is available again.
             </div>
-            <LButton
-                @click="deleteLocalData"
-                data-test="deleteLocalDatabase"
-                :disabled="!isConnected"
-            >
+            <LButton @click="deleteLocalData" data-test="deleteLocalDatabase">
                 Delete local cache
             </LButton>
         </LCard>
